pyproject.toml
[project]
name = "aiagenttutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "chromadb>=1.0.15",
    "generativeai>=0.0.1",
    "google-genai>=1.26.0",
    "ollama>=0.5.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "vertexai>=1.71.1",
]
